ai-language-tutor/v7/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
httpx
pydantic
aiosqlite
dspy